+++ trunk/cpsycle/audio/src/sampleiterator.h
@@ -91,7 +91,7 @@ INLINE uint32_t psy_audio_sampleiterator_position(psy_audio_SampleIterator* self
 }
 
 // Set Current sample position 
-void psy_audio_sampleiterator_setposition(psy_audio_SampleIterator*, uint32_t value);
+void psy_audio_sampleiterator_setposition(psy_audio_SampleIterator*, uintptr_t value);
 
 // Current sample Speed
 INLINE int64_t psy_audio_sampleiterator_speed(psy_audio_SampleIterator* self)
